tests/wpt/mozilla/tests/mozilla/element_classList.html

<!-- Remove this and update WPT metadata once DOMTokenList.toggle (#3138) is implemented. -->
<!DOCTYPE html>
<html>
    <head>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script>
          test(function() {
            let div = document.createElement("div");
            div.className = "foo bar";

            let classList = div.classList;
            div.className = "";

            assert_equals(classList.item(0), null, "classList.item(0) must return null when all classes have been removed");
            assert_equals(classList[0], undefined, "classList[0] must be undefined when all classes have been removed");
          });
        </script>
    </head>
    <body>
    </body>
</html>
